% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery-meta.R
\name{bqAlterTable}
\alias{bqAlterTable}
\title{Executes ALTER TABLE DML command to set table option value}
\usage{
bqAlterTable(x, name, value)
}
\arguments{
\item{x}{bq_table object}

\item{name}{name of the option}

\item{value}{value of the option}
}
\description{
Executes ALTER TABLE DML command to set table option value
}
\seealso{
https://cloud.google.com/bigquery/docs/reference/standard-sql/data-definition-language#alter_table_set_options_statement
}
