% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery-helpers.R
\name{bqCheckUniqueness}
\alias{bqCheckUniqueness}
\title{Checks that there are no duplicates in a table's key(s)}
\usage{
bqCheckUniqueness(table, dataset = bqDefaultDataset(), ...)
}
\arguments{
\item{table}{name of the table}

\item{dataset}{name of the dataset}

\item{...}{comma separated list of key columns}
}
\value{
TRUE if the table has no duplicate rows when grouped by keys
}
\description{
Checks that there are no duplicates in a table's key(s)
}
