% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery-dataset.R
\name{bqCopyDatasetSchema}
\alias{bqCopyDatasetSchema}
\title{Copies all tables from one dataset to another
through schema definitions}
\usage{
bqCopyDatasetSchema(from, to)
}
\arguments{
\item{from}{dataset object created by bigrquery::bq_dataset()}

\item{to}{dataset object created by bigrquery::bq_dataset()}
}
\description{
Copies all tables from one dataset to another
through schema definitions
}
