% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery.R
\name{bqCopyTable}
\alias{bqCopyTable}
\title{Copies table}
\usage{
bqCopyTable(from, to)
}
\arguments{
\item{from}{name of the source table}

\item{to}{name of the desitnation table}
}
\value{
result of the exectuion
}
\description{
Copies table
}
