% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery-dataset.R
\name{bqCopyTableSchema}
\alias{bqCopyTableSchema}
\title{Copies table through schema defintion}
\usage{
bqCopyTableSchema(from, to)
}
\arguments{
\item{from}{dataset object created by bigrquery::bq_table()}

\item{to}{dataset object created by bigrquery::bq_table()}
}
\description{
Copies table through schema defintion
}
