% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery-helpers.R
\name{bqCountDuplicates}
\alias{bqCountDuplicates}
\title{Counts any duplicate rows in a table when grouped by key(s)}
\usage{
bqCountDuplicates(table, dataset = bqDefaultDataset(), keys)
}
\arguments{
\item{table}{name of the table}

\item{dataset}{name of the dataset}

\item{keys}{vector of column names}
}
\value{
Total number of duplicate rows when grouped by keys
}
\description{
Counts any duplicate rows in a table when grouped by key(s)
}
