% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery.R
\name{bqCreatePartitionTable}
\alias{bqCreatePartitionTable}
\title{Creates partition table for a given sql}
\usage{
bqCreatePartitionTable(table, ga.properties, sql = NULL, file = NULL,
  existing.dates = NULL, missing.dates = NULL)
}
\arguments{
\item{table}{name of the destination table}

\item{ga.properties}{list of Google Analytics properties to populate table for}

\item{sql}{sql to use a source of the data}

\item{file}{if sql is not provided it will be read from the file}

\item{existing.dates}{dates that should be skipped}

\item{missing.dates}{dates calculation for which will be enforced}
}
\description{
Creates partition table for a given sql
}

