% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery-execute.R
\name{bqCreateTable}
\alias{bqCreateTable}
\title{Creates table using given sql}
\usage{
bqCreateTable(sql, table, dataset = bqDefaultDataset(),
  write.disposition = "WRITE_APPEND", priority = "INTERACTIVE",
  use.legacy.sql = bqUseLegacySql(), schema.file = NULL)
}
\arguments{
\item{sql}{string with sql statement or query template}

\item{table}{name of a table to be created}

\item{dataset}{name of the destination dataset}

\item{write.disposition}{defines whether records will be appended}

\item{priority}{sets priority of job execution to INTERACTIVE or BATCH}

\item{use.legacy.sql}{switches SQL dialect.
Defaults to value set in `BIGQUERY_LEGACY_SQL` env.}
}
\value{
results of the execution as returned by bigrquery::query_exec
}
\description{
Creates table using given sql
}
