% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery.R
\name{bqDataset}
\alias{bqDataset}
\alias{bqDatasetExists}
\alias{bqDataset}
\alias{bqCreateDataset}
\alias{bqDataset}
\alias{bqDeleteDataset}
\title{Checks if dataset exists in the project}
\usage{
bqDatasetExists(dataset = bqDefaultDataset(), project = bqDefaultProject())

bqCreateDataset(dataset = bqDefaultDataset(), project = bqDefaultProject())

bqDeleteDataset(dataset = bqDefaultDataset(), project = bqDefaultProject())
}
\arguments{
\item{dataset}{name of the dataset}

\item{project}{name of the project}
}
\value{
`bqDatasetExists` TRUE if dataset exists
}
\description{
Checks if dataset exists in the project
}
