% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery.R
\name{bqDatasetExists}
\alias{bqDatasetExists}
\title{Checks if dataset exists in the project}
\usage{
bqDatasetExists(dataset, project = Sys.getenv("BIGQUERY_PROJECT"))
}
\arguments{
\item{name}{of the dataset}
}
\value{
TRUE if dataset exists
}
\description{
Checks if dataset exists in the project
}
