% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery-partition.R
\name{bqDeletePartition}
\alias{bqDeletePartition}
\title{Deletes a specific partition of a partition table.}
\usage{
bqDeletePartition(table, date)
}
\arguments{
\item{table}{Name of the table}

\item{date}{Partition date}
}
\description{
Deletes a specific partition of a partition table.
}
