% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery.R
\name{bqExecuteFile}
\alias{bqExecuteFile}
\title{Gets data for a given file that contains SQL statement}
\usage{
bqExecuteFile(file, ...)
}
\arguments{
\item{file}{file with sql statment}

\item{...}{any parameters that will be used to fill in placeholders with sprintf}
}
\value{
results of execution as data.frame
}
\description{
Gets data for a given file that contains SQL statement
}

