% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery.gs.R
\name{bqExtractTable}
\alias{bqExtractTable}
\title{Exports BigQuery table into Google Cloud Storage file}
\usage{
bqExtractTable(
  table,
  dataset = bqDefaultDataset(),
  format = "CSV",
  compression = "GZIP"
)
}
\arguments{
\item{table}{name of the table to extract}

\item{dataset}{name of the dataset}

\item{format}{The exported file format. Possible values
include "CSV", "NEWLINE_DELIMITED_JSON" and "AVRO". Tables with nested or
repeated fields cannot be exported as CSV.}

\item{compression}{The compression type to use for exported files. Possible
values include "GZIP", "DEFLATE", "SNAPPY", and "NONE". "DEFLATE" and
"SNAPPY" are only supported for Avro.}
}
\value{
object of `bq_job`
}
\description{
Exports BigQuery table into Google Cloud Storage file
}
\seealso{
?bigrquery::bq_table_save
}
