% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery.R
\name{bqGetColumnNames}
\alias{bqGetColumnNames}
\title{Gets list of the column names for a given table}
\usage{
bqGetColumnNames(table)
}
\arguments{
\item{table}{name of the table}
}
\description{
Gets list of the column names for a given table
}
