% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery.R
\name{bqGetData}
\alias{bqGetData}
\title{Gets data for a given SQL statement or file that contains SQL}
\usage{
bqGetData(sql = NULL, file = NULL, ...)
}
\arguments{
\item{sql}{sql statement}

\item{file}{file with sql statment}

\item{...}{any parameters that will be used to fill in placeholders with sprintf}
}
\value{
results of execution as data.frame
}
\description{
Gets data for a given SQL statement or file that contains SQL
}
