% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery.R
\name{bqInitiateTable}
\alias{bqInitiateTable}
\title{Creates table from the json schema file.}
\usage{
bqInitiateTable(table, schema.file, partition = FALSE,
  dataset = Sys.getenv("BIGQUERY_DATASET"))
}
\arguments{
\item{table}{name of a table to be created}

\item{schema.file}{path to file with the table schema}

\item{partition}{time partitioned table will be created if set to TRUE}

\item{dataset}{name of the destination dataset}
}
\description{
Creates table from the json schema file.
}
\seealso{
https://cloud.google.com/bigquery/docs/reference/rest/v2/tables#schema.fields
}
