% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery.R
\name{bqInsertData}
\alias{bqInsertData}
\title{Inserts data into BigQuery table}
\usage{
bqInsertData(table, data, bq.project = Sys.getenv("BIGQUERY_PROJECT"),
  bq.dataset = Sys.getenv("BIGQUERY_DATASET"), append = TRUE,
  job.name = NULL, increment.field = NULL)
}
\arguments{
\item{table}{name of the target table}

\item{data}{data to be inserted}

\item{bq.project}{name of the destination project}

\item{bq.dataset}{name of the destination dataset}

\item{append}{specifies if data should be appended or truncated}

\item{job.name}{name of the ETL job that will be written to the metadata execution log}

\item{increment.field}{specifies field that is used for incremental data loads}
}
\value{
results of execution
}
\description{
Inserts data into BigQuery table
}
