% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery.R
\name{bqInsertLargeData}
\alias{bqInsertLargeData}
\title{Inserts LARGE data into BigQuery table}
\usage{
bqInsertLargeData(
  table,
  data,
  dataset = bqDefaultDataset(),
  chunks = NA_integer_,
  append = TRUE,
  schema.file = NULL,
  bucket = Sys.getenv("GCS_BUCKET")
)
}
\arguments{
\item{table}{name of the target table}

\item{data}{data to be inserted}

\item{dataset}{name of the destination dataset}

\item{chunks}{deprecated}

\item{append}{specifies if data should be appended or truncated}

\item{schema.file}{sets path to schema file for initialisation}

\item{bucket}{name of the GCS bucket where data will be saved in temp file.
Defaults to value of environment variable `GCS_BUCKET`}
}
\value{
results of execution
}
\description{
Truncates the target table and appends data broken in chunks
}
