% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery-partition.R
\name{bqInsertPartition}
\alias{bqInsertPartition}
\title{Inserts data table into a specific partition of a partition table.}
\usage{
bqInsertPartition(table, date, data, append = FALSE)
}
\arguments{
\item{table}{Name of the table}

\item{date}{Partition date}

\item{data}{Data table to insert}

\item{append}{Append to the partition if TRUE else overwrite}
}
\description{
Inserts data table into a specific partition of a partition table.
}
