% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery-dataset.R
\name{bqMatchDatasets}
\alias{bqMatchDatasets}
\title{Function to get list of datasets matching given pattern}
\usage{
bqMatchDatasets(pattern)
}
\arguments{
\item{pattern}{regex to match}
}
\value{
list of dataset objects (bigrquery)
}
\description{
Function to get list of datasets matching given pattern
}
