% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery.R
\name{bqTransformPartition}
\alias{bqTransformPartition}
\alias{bqRefreshPartitionData}
\title{Functions to transforms partitioned data form one table to another}
\usage{
bqTransformPartition(table, file, ..., priority = "INTERACTIVE")

bqRefreshPartitionData(table, file, ..., priority = "BATCH")
}
\arguments{
\item{table}{destination partition table where results of the query will be saved}

\item{file}{path to the sql file that will be used for the transformation}

\item{...}{parameters that will be passed via `sprintf` to build dynamic SQL.
partition date will be always passed first in format `yyyymmdd`
followed by arguments in `...`}

\item{priority}{sets priority of job execution to INTERACTIVE or BATCH}
}
\description{
`bqTransformPartition` creates new partitions for the missing dates

`bqRefreshPartitionData` updates existing partitions in the target table
}
