% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery.R
\name{bqPartitionName}
\alias{bqPartitionName}
\title{Creates partition name by combining table and partition date.}
\usage{
bqPartitionName(table, date)
}
\arguments{
\item{table}{Name of the table}

\item{date}{Partition date}
}
\value{
Full partition table name
}
\description{
Creates partition name by combining table and partition date.
}
