% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery.R
\name{bqProject}
\alias{bqProject}
\alias{bqProjectDatasets}
\alias{bqProjectTables}
\title{Functions to work with BigQuery projects}
\usage{
bqProjectDatasets(project = bqDefaultProject())

bqProjectTables(project = bqDefaultProject())
}
\arguments{
\item{project}{name of the bigquery project}
}
\description{
Family of functions for common operations on projects

Gets list of datasets for a given project

Gets metadata of all tables of a project into a data.table
}
\details{
Function queries __TABLES__ table for each dataset in the project
}
