% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery.R
\name{bqQuery}
\alias{bqQuery}
\alias{bqExecuteFile}
\alias{bqExecuteQuery}
\alias{bqExecuteSql}
\title{Functions that execute query against BigQuery database}
\usage{
bqExecuteFile(file, ...)

bqExecuteQuery(query, ...)

bqExecuteSql(sql, ..., use.legacy.sql = bqUseLegacySql())
}
\arguments{
\item{file}{file with sql statement}

\item{...}{any parameters that will be used to replace placeholders in the query template}

\item{query}{template of the sql statement}

\item{sql}{string with sql statement or query template}

\item{use.legacy.sql}{switches SQL dialect.
Defaults to value set in `BIGQUERY_LEGACY_SQL` env.}
}
\description{
Execute templated query given as string or file.
Placeholders in template are replaced with values provided in ellipsis parameter with sprintf
}
