% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery-partition.R
\name{bqRefreshPartitionData}
\alias{bqRefreshPartitionData}
\title{Functions to update existing partitions in the target table}
\usage{
bqRefreshPartitionData(table, file, ..., priority = "BATCH",
  use.legacy.sql = bqUseLegacySql())
}
\arguments{
\item{table}{destination partition table where results of the query will be saved}

\item{file}{path to the sql file that will be used for the transformation}

\item{...}{parameters that will be passed via `sprintf` to build dynamic SQL.
partition date will be always passed first in format `yyyymmdd`
followed by arguments in `...`}

\item{priority}{sets priority of job execution to INTERACTIVE or BATCH}

\item{use.legacy.sql}{sets SQL flavour}
}
\description{
`bqRefreshPartitionData` updates existing partitions in the target table
}
