% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery.R
\name{bqSaveSchema}
\alias{bqSaveSchema}
\alias{bqExtractSchema}
\title{Save file inferred from data.table to JSON file}
\usage{
bqSaveSchema(dt, file)

bqExtractSchema(dt)
}
\arguments{
\item{dt}{data.table, table name or bq_fields from which schema will be extracted into JSON}

\item{file}{path to the schema file}
}
\description{
Save file inferred from data.table to JSON file

Extract schema as JSON from data.table
}
