% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery-schema.R
\name{bqSetdiffSchemas}
\alias{bqSetdiffSchemas}
\title{Finds asymmetric difference between two sets of bq_fields objects}
\usage{
bqSetdiffSchemas(x, y)
}
\arguments{
\item{x}{bq_fields object}

\item{y}{bq_fields object}
}
\description{
Finds asymmetric difference between two sets of bq_fields objects
}
\seealso{
setdiff bqMatchField
}
