% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery.R
\name{bqTable}
\alias{bqTable}
\alias{bqTable}
\alias{bqTableExists}
\alias{bqTable}
\alias{bqDeleteTable}
\alias{bqTable}
\alias{bqInitiateTable}
\alias{bqTable}
\alias{bqTableSchema}
\title{Functions to work with BigQuery tables}
\usage{
bqTableExists(table, dataset = bqDefaultDataset())

bqDeleteTable(table, dataset = bqDefaultDataset())

bqInitiateTable(table, schema.file, partition = FALSE,
  dataset = bqDefaultDataset())

bqTableSchema(table, dataset = bqDefaultDataset())
}
\arguments{
\item{table}{name of the table}

\item{dataset}{name of the dataset}

\item{schema.file}{path to file with the table schema}

\item{partition}{time partitioned table will be created if set to TRUE}
}
\value{
`bqTableExists` TRUE if table exists

`bqDeleteTable` TRUE if table was deleted
}
\description{
Family of functions for common operations on tables
}
\seealso{
https://cloud.google.com/bigquery/docs/reference/rest/v2/tables#schema.fields
}
