% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery.R
\name{bqTableExists}
\alias{bqTableExists}
\title{Checks if table exists}
\usage{
bqTableExists(table.name)
}
\arguments{
\item{table.name}{name of the table}
}
\value{
TRUE if table exists
}
\description{
Checks if table exists
}
