% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery.R
\name{bqTransformPartition}
\alias{bqTransformPartition}
\title{Transforms data from one partition table to another partition table}
\usage{
bqTransformPartition(table, file, ...)
}
\arguments{
\item{table}{destination partition table where resutls of the query will be saved}
}
\description{
Transforms data from one partition table to another partition table
}
