% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery-meta.R
\name{bqUpdateTableDescription}
\alias{bqUpdateTableDescription}
\title{Sets description of the table}
\usage{
bqUpdateTableDescription(
  table,
  description,
  dataset = bqDefaultDataset(),
  project = bqDefaultProject()
)
}
\arguments{
\item{table}{name of the table to update}

\item{description}{character scalar with table description}

\item{dataset}{name of the dataset}

\item{project}{name of the project}
}
\description{
Sets description of the table
}
