% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{conformHeader}
\alias{conformHeader}
\title{Replace " ", "-" and "_" with "." in the header.}
\usage{
conformHeader(dt)
}
\arguments{
\item{dt}{data.table to correc the header for.}
}
\description{
Replace " ", "-" and "_" with "." in the header.
}
