% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{convert_cols}
\alias{convert_cols}
\alias{convert_cols.data.table}
\title{convert all columns of specified type to another type}
\usage{
convert_cols(x, ...)

\method{convert_cols}{data.table}(x, from_class, to_class, inplace = TRUE)
}
\arguments{
\item{x}{an input data.table}

\item{from_class}{original type/class}

\item{to_class}{target type/class}
}
\description{
convert all columns of specified type to another type
}
\details{
Modifies the referenced data.tables!
}
