% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery.R
\name{createPartitionTable}
\alias{createPartitionTable}
\title{Creates partition in specified table in BigQuery.}
\usage{
createPartitionTable(table, sql = NULL, file = NULL,
  existing.dates = NULL, missing.dates = NULL)
}
\arguments{
\item{table}{name of the new table}

\item{sql}{source for the table as string.}

\item{file}{source for the tabel as file.}
}
\description{
Creates partition in specified table in BigQuery.
}
\note{
sql or file must be provided.
}

