% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery.R
\name{createRangeTable}
\alias{createRangeTable}
\title{Creates range teable in BigQuery.}
\usage{
createRangeTable(table, sql = NULL, file = NULL)
}
\arguments{
\item{table}{name of the new table}

\item{sql}{source for the table as string.}

\item{file}{source for the tabel as file.}
}
\description{
Creates range teable in BigQuery.
}
\note{
sql or file must be provided.
}

