% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{Date2Char}
\alias{Date2Char}
\title{Convert all Date and POSIXct columns to character columns}
\usage{
Date2Char(x, ..., inplace = TRUE)
}
\arguments{
\item{x}{a data.table object}

\item{...}{args passed to `as.Date()`}

\item{inplace}{Should x be modifed by reference? If FALSE a modified copy is returned.}
}
\description{
Convert all Date and POSIXct columns to character columns
}
