% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgresql.R
\name{dbExecuteQuery}
\alias{dbExecuteQuery}
\title{Executes SQL query agains the database}
\usage{
dbExecuteQuery(sql, ...)
}
\arguments{
\item{sql}{string with sql query}

\item{...}{any parameters that will be used to fill in placeholders with sprintf}
}
\description{
Executes SQL query agains the database
}
