% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgresql.R
\name{dbExecuteQueryFile}
\alias{dbExecuteQueryFile}
\title{Executes statement in the SQL file}
\usage{
dbExecuteQueryFile(sql.file)
}
\arguments{
\item{sql.file}{path to the sql file}
}
\description{
Executes statement in the SQL file
}

