% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgresql.R
\name{dbGetConnection}
\alias{dbGetConnection}
\title{Creates db connection based on the environment variables}
\usage{
dbGetConnection()
}
\description{
Creates db connection based on the environment variables
}
