% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclick.R
\name{dcPredictionBody}
\alias{dcPredictionBody}
\title{Creates list for the DoubleClick API call from prediction variables}
\usage{
dcPredictionBody(
  clickId,
  conversionId,
  datetime,
  custom.metrics,
  revenue,
  currency.code
)
}
\arguments{
\item{clickId}{gclid for the DoubleClick match for the session}

\item{conversionId}{unique identifier for the conversion}

\item{datetime}{POSIX timestamp in milliseconds}

\item{custom.metrics}{list with named vectors of custom metric values}

\item{revenue}{data to be loaded to Revenue}

\item{currency.code}{currency associated to the Revenue}
}
\description{
Creates list for the DoubleClick API call from prediction variables
}
