% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclick.R
\name{dcUpdateAvailability}
\alias{dcUpdateAvailability}
\title{Updates availability of data in SA360 platform to the current time}
\usage{
dcUpdateAvailability(
  agency.id,
  advertiser.id,
  segmentation.type = "FLOODLIGHT",
  segmentation.id,
  segmentation.name
)
}
\arguments{
\item{agency.id}{agencyId, e.g. `12300000000000456`}

\item{advertiser.id}{advertiserId e.g. `45600000000010291`}

\item{segmentation.type}{segmentationType e.g. `FLOODLIGHT`}

\item{segmentation.id}{segmentationId e.g. `25700000001464142`}

\item{segmentationName}{segmentationName e.g. `Test`}
}
\description{
Updates availability of data in SA360 platform to the current time
}
\seealso{
https://developers.google.com/search-ads/v2/how-tos/conversions/offline-conversions#send-a-conversion.updateavailability-request
}
