% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclick.R
\name{dcWriteCustomMetics}
\alias{dcWriteCustomMetics}
\title{Writes custom metrics to DoubleClick Floodlight
requires gar_auth()}
\usage{
dcWriteCustomMetics(clickId, conversionId, metrics)
}
\arguments{
\item{clickId}{gclid unique identifier of the google click}

\item{conversionId}{unique identifier of the conversion}

\item{metrics}{named vector of custom metrics}
}
\description{
Writes custom metrics to DoubleClick Floodlight
requires gar_auth()
}
