% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclick.R
\name{dcWriteCustomMetrics}
\alias{dcWriteCustomMetrics}
\title{Writes custom metrics to DoubleClick Floodlight
requires gar_auth()}
\usage{
dcWriteCustomMetrics(
  clickId,
  conversionId,
  timestamp,
  metrics,
  revenue,
  currency.code
)
}
\arguments{
\item{clickId}{gclid unique identifier of the google click}

\item{conversionId}{unique identifier of the conversion}

\item{timestamp}{POSIX timestamp in seconds as integer}

\item{metrics}{named vector of custom metrics}

\item{revenue}{data to be loaded to Revenue}

\item{currency.code}{currency associated to the Revenue}
}
\value{
response from the call to DoubleClick search API.
}
\description{
Writes custom metrics to DoubleClick Floodlight
requires gar_auth()
}
