% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{disaggregate}
\alias{disaggregate}
\title{Splits table rows into multiple by the integer "by". All other fields are duplicated.}
\usage{
disaggregate(dt, by)
}
\arguments{
\item{dt}{data.table to disaggregate.}

\item{by}{integer by which each line is split into units of 1.}
}
\description{
Splits table rows into multiple by the integer "by". All other fields are duplicated.
}
