% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt.R
\name{dtq}
\alias{dtq}
\title{data.table query - alias for data.table's \code{`[`} method}
\usage{
dtq(x, i, j, by, keyby, with = TRUE,
  nomatch = getOption("datatable.nomatch"), mult = "all",
  roll = FALSE, rollends = if (roll == "nearest") c(TRUE, TRUE) else if
  (roll >= 0) c(FALSE, TRUE) else c(TRUE, FALSE), which = FALSE, .SDcols,
  verbose = getOption("datatable.verbose"),
  allow.cartesian = getOption("datatable.allow.cartesian"),
  drop = NULL, on = NULL)
}
\arguments{
\item{.}{see \code{\link[data.table]{data.table}}}
}
\description{
data.table query - alias for data.table's \code{`[`} method
}
\details{
see \code{\link[data.table]{data.table}}
}
