% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipes.R
\name{etl_write.character}
\alias{etl_write.character}
\title{Write tables to CSV files}
\usage{
\method{etl_write}{character}(to, x, name = NULL, schema = NULL, ...)
}
\arguments{
\item{to}{name of the target file}

\item{x}{data.table}

\item{name}{ignored}

\item{schema}{ignored}

\item{...}{args passed to `fwrite()`}
}
\description{
Write tables to CSV files
}
