% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipes.R
\name{etl_write.DBIConnection}
\alias{etl_write.DBIConnection}
\title{Write tables via DBI connection}
\usage{
\method{etl_write}{DBIConnection}(to, x, name, schema = NULL, ...)
}
\arguments{
\item{to}{}

\item{x}{data.table}

\item{name}{name of target table}

\item{schema}{prefix to table name}

\item{...}{args passed to dbWriteTable}
}
\description{
Write tables via DBI connection
}
