% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipes.R
\name{etl_write.environment}
\alias{etl_write.environment}
\title{Write tables to R environment}
\usage{
\method{etl_write}{environment}(to, x, name, ...)
}
\arguments{
\item{to}{R environemnt}

\item{x}{data.table}

\item{name}{name of the target object}

\item{...}{ignored}
}
\description{
Write tables to R environment
}
