% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipes.R
\name{etl_write}
\alias{etl_write}
\title{Write (or load) part of ETL}
\usage{
etl_write(to, x, name, schema = NULL, ...)
}
\arguments{
\item{to}{DBI/R environment/file name}

\item{x}{data.frame object}

\item{name}{character;target name}

\item{schema}{character;}

\item{...}{arguments passed to write/load function}
}
\description{
Write (or load) part of ETL
}
