% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{etlAddJob}
\alias{etlAddJob}
\title{Adds ETL job}
\usage{
etlAddJob(job, increment.name, increment.type)
}
\arguments{
\item{job}{name of the job}

\item{increment.name}{name of the increment}

\item{increment.type}{type of the increment}
}
\description{
Adds ETL job
}

