% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{etlGetIncrement}
\alias{etlGetIncrement}
\title{Gets latest increment for the}
\usage{
etlGetIncrement(job)
}
\arguments{
\item{job}{name of the job}
}
\value{
increment value for the last execution of the job
}
\description{
Gets latest increment for the
}

