% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{etlGetIncrementType}
\alias{etlGetIncrementType}
\title{Gets increment value from}
\usage{
etlGetIncrementType(job)
}
\arguments{
\item{job}{name of the etl job}
}
\description{
Gets increment value from
}

