% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{etlJobExists}
\alias{etlJobExists}
\title{Checks if ETL job exists in the etl_jobs table}
\usage{
etlJobExists(job)
}
\arguments{
\item{job}{name of the ETL job}
}
\description{
Checks if ETL job exists in the etl_jobs table
}
