% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{etlLogExecution}
\alias{etlLogExecution}
\title{Logs execution of the ETL job}
\usage{
etlLogExecution(job, increment.value, records = 0)
}
\arguments{
\item{job}{name of the etl job}

\item{increment.value}{maximum value in the field that used for increment}

\item{records}{number of records processed by execution}
}
\description{
Logs execution of the ETL job
}
