% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{excel_worksheets}
\alias{excel_worksheets}
\title{Reads all sheets from an excel xlsx file into a list}
\usage{
excel_worksheets(file, ...)
}
\arguments{
\item{file}{file name}

\item{...}{args passed to openxlsx::readWorkbook}
}
\value{
a list of data.frames
}
\description{
Reads all sheets from an excel xlsx file into a list
}
