% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{excel_worksheets_readxl}
\alias{excel_worksheets_readxl}
\title{Reads all sheets from an excel xlsx file into a list}
\usage{
excel_worksheets_readxl(file)
}
\arguments{
\item{file}{file name}
}
\description{
Reads all sheets from an excel xlsx file into a list
}
