% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery.R
\name{extensionFromFormat}
\alias{extensionFromFormat}
\title{Converts possible BigQuery extract formats to file extensions}
\usage{
extensionFromFormat(format, compression = "NONE")
}
\arguments{
\item{format}{format of the file}

\item{compression}{level of compression to apply to the output}
}
\value{
corresponding file extension
}
\description{
Converts possible BigQuery extract formats to file extensions
}
