% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftp.R
\name{ftp-url}
\alias{ftp-url}
\alias{ftpBaseUrl}
\alias{ftpFullUrl}
\alias{ftpListFiles}
\title{Base url to FTP based on the environment variables}
\usage{
ftpBaseUrl(storage = "ftps")

ftpFullUrl(path)

ftpListFiles(path, pattern, verbose = FALSE, level = 0)
}
\arguments{
\item{storage}{- defines whether files will be read from FTP or FTPS. Defaults to `ftps`.}

\item{path}{relative or full path to the FTP object or folder}

\item{pattern}{regex pattern for file name filter}

\item{verbose}{defines whether FTP curl feedback will be printed}

\item{level}{internal parameter to track subfolder levels}
}
\description{
FTP(s) path is created for specified storage and environment variables

Gets full URL to the FTP object

Recursively scans given FTP folder for files that match specified regex pattern
}
\details{
Environment variables required for FTP(S) access:
 * `FTP_USER` - User name
 * `FTP_PASSWORD` = Password
 * `FTP_SERVER` = FTP domain
}
