% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery.R
\name{gaGetShop}
\alias{gaGetShop}
\title{Gets the shop code from the GA properties vector.}
\usage{
gaGetShop(ga.properties, property)
}
\arguments{
\item{ga.properties}{named vector of Google Analytics properties.
Names are ISO2 codes of the country.}

\item{property}{number of a property in Google Analytics.}
}
\value{
gets site code for a given propertiy code
}
\description{
Gets the shop code from the GA properties vector.
}

