% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery.R
\name{getExistingDates}
\alias{getExistingDates}
\title{Gets existing dates from wildcard tables in BigQuery}
\usage{
getExistingDates(dataset, table.prefix)
}
\arguments{
\item{dataset}{name of a dataset}

\item{table.prefix}{name of the table before the wildcard}
}
\value{
string vector of dates
}
\description{
Gets existing dates from wildcard tables in BigQuery
}
