% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery.R
\name{getExistingPartitionDates}
\alias{getExistingPartitionDates}
\title{Gets existing dates for date partitioned table in BigQuery}
\usage{
getExistingPartitionDates(table)
}
\arguments{
\item{table}{name of a table}
}
\value{
string vector of dates
}
\description{
Gets existing dates for date partitioned table in BigQuery
}

