% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery.R
\name{getMissingDates}
\alias{getMissingDates}
\title{Gets dates that are missing from the date range for a give list of existing dates}
\usage{
getMissingDates(start.date, end.date, existing.dates)
}
\arguments{
\item{start.date}{begining of the period}

\item{end.date}{end of the period}

\item{existing.dates}{vector of existing dates that should be exlcuded}
}
\value{
vector of dates from the period that don't exist in the give vector
}
\description{
Gets dates that are missing from the date range for a give list of existing dates
}

