% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlespreadsheets.R
\name{gsAuth}
\alias{gsAuth}
\title{Authentication for google sheets. Requires access_token.json path as env. var.}
\usage{
gsAuth()
}
\description{
Authentication for google sheets. Requires access_token.json path as env. var.
}
