% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlespreadsheets.R
\name{gsLoadAll}
\alias{gsLoadAll}
\title{Loads all sheets from a google spreadsheet into a list with the tab name as the list element name.}
\usage{
gsLoadAll(key)
}
\arguments{
\item{key}{Something that identifies a Google Sheet: its file ID, a URL from which we can recover the ID, an instance of googlesheets4_spreadsheet (returned by gs4_get()), or a dribble, which is how googledrive represents Drive files. Processed through as_sheets_id().}
}
\description{
Loads all sheets from a google spreadsheet into a list with the tab name as the list element name.
}
