% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery.gs.R
\name{gsPathUri}
\alias{gsPathUri}
\title{makes full path to the gs file from relative part}
\usage{
gsPathUri(path)
}
\arguments{
\item{path}{relative path to the gs object}
}
\description{
makes full path to the gs file from relative part
}
