% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery.gs.R
\name{gsTablePath}
\alias{gsTablePath}
\title{makes relative path to a file to mirror table path}
\usage{
gsTablePath(x, format, compression)
}
\description{
makes relative path to a file to mirror table path
}
