% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery.gs.R
\name{gsUri}
\alias{gsUri}
\title{Gets gs uri to mirror table by name}
\usage{
gsUri(x, format, compression)
}
\arguments{
\item{x}{bq_table object}

\item{format}{format of the file}

\item{compression}{compression applied to the filed}
}
\description{
Gets gs uri to mirror table by name
}
