% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influx.R
\name{influxGetConnection}
\alias{influxGetConnection}
\title{Creates db connection based on the environmental variables}
\usage{
influxGetConnection(scheme = "http", host = Sys.getenv("INFLUX_HOST"),
  user = Sys.getenv("INFLUX_USER"), port = 8086,
  pwd = Sys.getenv("INFLUX_PASSWORD"))
}
\arguments{
\item{scheme}{look at influx_connection}

\item{host}{look at influx_connection}

\item{user}{look at influx_connection}

\item{pwd}{look at influx_connection}

\item{host}{look at influx_connection}
}
\description{
Creates db connection based on the environmental variables
}
