% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influx.R
\name{influxLog}
\alias{influxLog}
\title{Logs the result of a job in the influx metric table}
\usage{
influxLog(con = influxConnection(), db = Sys.getenv("INFLUX_DB"), job, val,
  metric = Sys.getenv("INFLUX_METRIC"), env = Sys.getenv("INFLUX_PROJECT"))
}
\arguments{
\item{con}{a connection object.
defaults to connection that was set via `influxConnection()` call}

\item{db}{the database to write, defaults to value in `INFLUX_DB` env var}

\item{job}{the name of the job to log}

\item{val}{the variable to log}

\item{metric}{the metric to be written to, defaults to value in `INFLUX_METRIC` env var}

\item{env}{the environment to write test/prod, defaults to value in `INFLUX_PROJECT` env var}
}
\description{
Logs the result of a job in the influx metric table
}
