% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{int64_to_numeric}
\alias{int64_to_numeric}
\title{Convert all integer64 columns to numeric}
\usage{
int64_to_numeric(x, ..., inplace = TRUE)
}
\arguments{
\item{x}{a data.table object}

\item{...}{args passed to `as.numeric()`}

\item{inplace}{Should x be modifed by reference? If FALSE a modified copy is returned.}
}
\description{
Convert all integer64 columns to numeric
}
