% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{is_apply}
\alias{is_apply}
\title{is function vectorized over list of objects (not classes)}
\usage{
is_apply(x, class_name)
}
\arguments{
\item{x}{list of objects}

\item{class_name}{name of the class}
}
\description{
is function vectorized over list of objects (not classes)
}
