% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{is_apply2}
\alias{is_apply2}
\title{is function vectorized over list of objects and vector of classes}
\usage{
is_apply2(x, class_names)
}
\arguments{
\item{x}{list of objects}

\item{class_names}{name of the class}
}
\description{
is function vectorized over list of objects and vector of classes
}
\details{
For each object in the list, it is checked againts all the class names given. If any of the test is successfull it returns TRUE for given object.
}
