% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery-mock.R
\name{make_bq_dataset_query_stub}
\alias{make_bq_dataset_query_stub}
\title{Creates a function that wraps a given query into a query with stub data}
\usage{
make_bq_dataset_query_stub(data.query.path)
}
\arguments{
\item{data.query.path}{path to the file with stub data query}
}
\description{
Creates a function that wraps a given query into a query with stub data
}
