% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclick.R
\name{metricsToList}
\alias{metricsToList}
\title{Converts named vector of metrics to a list}
\usage{
metricsToList(metrics)
}
\arguments{
\item{metrics}{named vector with custom metrics values}
}
\description{
Converts named vector of metrics to a list
}
