% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipes.R
\name{pipe_table}
\alias{pipe_table}
\alias{pipe_tables}
\alias{etl_pipe}
\title{complete ETL pipe in a single function}
\usage{
pipe_table(
  name,
  from,
  to,
  name_target = NULL,
  transform = NULL,
  from_schema = NULL,
  to_schema = NULL,
  read_args = list(),
  write_args = list(overwrite = TRUE),
  asDT = TRUE,
  lowercase = FALSE
)

pipe_tables(tables, ...)

etl_pipe(
  from,
  to,
  transform = NULL,
  read_args = list(),
  write_args = list(overwrite = TRUE),
  asDT = TRUE,
  lowercase = FALSE
)
}
\arguments{
\item{name}{name of the table}

\item{from}{DBI/R environment/file name/...}

\item{to}{DBI/R environment/file name/...}

\item{name_target}{optional parameter specifying new name of the table at the target}

\item{transform}{transformation function}

\item{from_schema}{optional parameter with name of schema at the origin database}

\item{to_schema}{optional parameter with name of schema at the target database}

\item{read_args}{list of arguments passed to read/extract function}

\item{write_args}{list of arguments passed to write/load function}

\item{asDT}{logical; convert the data.frame to data.table during the etl?}

\item{lowercase}{logical; lowercase all column names?}

\item{tables}{vector of table names}

\item{...}{args passed to `pipe_table()`}
}
\description{
complete ETL pipe in a single function

pipe_table vectorized

complete ETL pipe
}
