% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-helpers.R
\name{readSqlGlue}
\alias{readSqlGlue}
\title{Glues variables passed in ellipsis into a text of the file}
\usage{
readSqlGlue(file, ...)
}
\arguments{
\item{file}{file with text to glue}

\item{...}{named parameters for glue}
}
\value{
text with values replaced based the template
}
\description{
Glues variables passed in ellipsis into a text of the file
}
