% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws-s3.R
\name{s3GetData}
\alias{s3GetData}
\title{Loads data from several files in S3 based on the path prefix}
\usage{
s3GetData(path, header = T, bucket = s3DefaultBucket(),
  root = s3DefaultRoot(), s3Get.FUN = s3GetFile, fill = T)
}
\arguments{
\item{path}{is the path to the S3 object}

\item{header}{flag defines whether file has header}

\item{bucket}{name of the S3 bucket.
Defaults to value in \code{AWS_S3_BUCKET} environment variable.}

\item{root}{project root path that is appended before the path in the argument.
Defaults to value in \code{AWS_S3_ROOT} environment variable.}

\item{s3Get.FUN}{Function that will be used to read data from the individual files.}

\item{fill}{see data.table::rbindList}
}
\value{
data.table with data combined from files. It will be Null data.table if path
did not match any of the files in S3 bucket.
}
\description{
Loads data from several files in S3 based on the path prefix
}
\note{
You need make sure that all targeted files have the same header signature:
  order of the fields should not matter.
}
