% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws-s3.R
\name{s3GetFile.gz}
\alias{s3GetFile.gz}
\title{Loads gz file from s3 and reads it as data.table from csv}
\usage{
s3GetFile.gz(path, bucket = Sys.getenv("AWS_S3_BUCKET"),
  root = Sys.getenv("AWS_S3_ROOT"))
}
\arguments{
\item{path}{S3 object path starting after root folder.}

\item{bucket}{name of the S3 bucket}

\item{root}{project root path that is appended before the path, e.g. "/prod/"}
}
\description{
Loads gz file from s3 and reads it as data.table from csv
}
