% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws-s3.R
\name{s3GetFile.json.gz}
\alias{s3GetFile.json.gz}
\title{Loads gzip-ed JSON file from S3 and reads it as data.table.}
\usage{
s3GetFile.json.gz(path, bucket = Sys.getenv("AWS_S3_BUCKET"),
  root = Sys.getenv("AWS_S3_ROOT"))
}
\arguments{
\item{path}{defines the path to the object post project root path.}

\item{bucket}{name of the S3 bucket}

\item{root}{project root path that is appended before the path, e.g. "/prod/"}
}
\description{
Loads gzip-ed JSON file from S3 and reads it as data.table.
}
