% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws-s3.R
\name{s3GetFile}
\alias{s3GetFile}
\alias{s3GetFile.gz}
\alias{s3GetFile.rds}
\alias{s3GetFile.zip}
\alias{s3GetFile.json.gz}
\title{Loads data from AWS S3 into data.table object}
\usage{
s3GetFile(path, header = T, bucket = Sys.getenv("AWS_S3_BUCKET"),
  root = Sys.getenv("AWS_S3_ROOT"))

s3GetFile.gz(path, bucket = Sys.getenv("AWS_S3_BUCKET"),
  root = Sys.getenv("AWS_S3_ROOT"))

s3GetFile.rds(path, bucket = Sys.getenv("AWS_S3_BUCKET"),
  root = Sys.getenv("AWS_S3_ROOT"))

s3GetFile.zip(path, bucket = Sys.getenv("AWS_S3_BUCKET"),
  root = Sys.getenv("AWS_S3_ROOT"))

s3GetFile.json.gz(path, bucket = Sys.getenv("AWS_S3_BUCKET"),
  root = Sys.getenv("AWS_S3_ROOT"))
}
\arguments{
\item{path}{is the path to the S3 object}

\item{header}{flag defines whether file has header}

\item{bucket}{name of the S3 bucket.
Defautls to value in `AWS_S3_BUCKET` environment variable.}

\item{root}{project root path that is appended before the path in the argument.
Defaults to value in `AWS_S3_ROOT` environment variable.}
}
\value{
`s3GetFile` gets data from source `.csv` file

`s3GetFile.gz` loads data from `.gz` files

`s3GetFile.rds` loads data from `.rds` files

`s3GetFile.zip` loads data from `.zip` files

`s3GetFile.json.gz` loads data from `.json.gz` files
}
\description{
Convention based wrapper functions that allow to load
data files in S3 into data.table
}
