% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws-s3.R
\name{s3GetFile.rds}
\alias{s3GetFile.rds}
\title{Saves file as rds to S3}
\usage{
s3GetFile.rds(path, bucket = Sys.getenv("AWS_S3_BUCKET"),
  root = Sys.getenv("AWS_S3_ROOT"))
}
\arguments{
\item{path}{is the path to the S3 object}

\item{bucket}{name of the S3 bucket}

\item{root}{project root path that is appended before the path}
}
\value{
data.table from the source csv file
}
\description{
Saves file as rds to S3
}
