% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws-s3.R
\name{s3GetFile.zip}
\alias{s3GetFile.zip}
\title{Loads zipped csv file from s3 and reads it as data.table}
\usage{
s3GetFile.zip(path, bucket = Sys.getenv("AWS_S3_BUCKET"),
  root = Sys.getenv("AWS_S3_ROOT"))
}
\arguments{
\item{path}{defines the path to the object post project root path.}

\item{bucket}{name of the S3 bucket}

\item{root}{project root path that is appended before the path, e.g. "/prod/"}
}
\description{
Loads zipped csv file from s3 and reads it as data.table
}
