% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws-s3.R
\name{s3GetObjectPath}
\alias{s3GetObjectPath}
\title{Gets the path to the object based on the root path setup via
environment variable}
\usage{
s3GetObjectPath(relative.path)
}
\arguments{
\item{relative.path}{defines the path to the object post project root path.}
}
\description{
Gets the path to the object based on the root path setup via
environment variable
}

