% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws-s3.R
\name{s3PutFile.gz}
\alias{s3PutFile.gz}
\title{Saves data.table as csv.gz to S3}
\usage{
s3PutFile.gz(dt, path, bucket = Sys.getenv("AWS_S3_BUCKET"),
  root = Sys.getenv("AWS_S3_ROOT"))
}
\arguments{
\item{dt}{data.table to save}

\item{path}{S3 object path starting after root folder.}

\item{bucket}{name of the S3 bucket}

\item{root}{project root path that is appended before the path, e.g. "/prod/"}
}
\description{
Saves data.table as csv.gz to S3
}
