% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws-s3.R
\name{s3PutFile.json.gz}
\alias{s3PutFile.json.gz}
\title{Saves data.table as gzip-ed JSON to S3.}
\usage{
s3PutFile.json.gz(dt, path, bucket = Sys.getenv("AWS_S3_BUCKET"),
  root = Sys.getenv("AWS_S3_ROOT"))
}
\arguments{
\item{dt}{data table to upload}

\item{path}{defines the path to the object post project root path.}

\item{bucket}{name of the S3 bucket}

\item{root}{project root path that is appended before the path, e.g. "/prod/"}
}
\description{
Saves data.table as gzip-ed JSON to S3.
}
