% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{safeLookup}
\alias{safeLookup}
\title{Function to safely append values to a data set from another data set via keys}
\usage{
safeLookup(data, lookup, by, select = setdiff(colnames(lookup), by))
}
\arguments{
\item{data}{data set to which fields will be appended}

\item{lookup}{data set from which fields will be appended}

\item{by}{fields that represent the keys}

\item{select}{fields taht will be appended}
}
\value{
data set with new fields
TODO @Daniel : add unit tests
}
\description{
Function to safely append values to a data set from another data set via keys
}
