% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-helpers.R
\name{sqlRangeLabel}
\alias{sqlRangeLabel}
\alias{sqlRangeIndex}
\title{Functions that allow to convert numeric vector in to SQL case statement}
\usage{
sqlRangeLabel(field, limits, labels = rangeLabel)

sqlRangeIndex(field, limits, labels = rangeIndex)
}
\arguments{
\item{field}{field name to be used for the binning}

\item{limits}{vector of separator values}

\item{labels}{function that turns index, low}
}
\description{
Creates CASE statement that turns numberic field into ranges

Creates CASE statement that turns index for the range field
}
