% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery-mock.R
\name{with_mock_bigquery}
\alias{with_mock_bigquery}
\title{Executes given query against stub data from a given query file}
\usage{
with_mock_bigquery(x, data.query.path)
}
\arguments{
\item{x}{query to execute}

\item{data.query.path}{path to a query with stub data}
}
\description{
Executes given query against stub data from a given query file
}
