% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{worksheets_to_csvs}
\alias{worksheets_to_csvs}
\title{Writes list of data.frames as csv files into a folder}
\usage{
worksheets_to_csvs(sheets, path, ...)
}
\arguments{
\item{sheets}{a list of data.frames}

\item{path}{path to a destination dir}

\item{...}{args passed to fwrite}
}
\description{
Writes list of data.frames as csv files into a folder
}
