% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{add_diff_value}
\alias{add_diff_value}
\title{Add a new column as a differenced value of existing column}
\usage{
add_diff_value(.data, value_var = NULL, by_vars = NULL, postfix = "_diff")
}
\arguments{
\item{.data}{input data.table}

\item{value_var}{character; name of the original column}

\item{by_vars}{character; vector of dimensions for 'group by'}

\item{postfix}{character; posfix is added at the end of the name of the new column}
}
\value{

}
\description{
Add a new column as a differenced value of existing column
}
