% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery-helpers.R
\name{bqAssertUnique}
\alias{bqAssertUnique}
\title{Counts any duplicate rows in a table when grouped by key(s)}
\usage{
bqAssertUnique(table, dataset = bqDefaultDataset(), keys)
}
\arguments{
\item{table}{name of the table}

\item{dataset}{name of the dataset}

\item{keys}{vector of column names}
}
\value{
Throws exception if duplicates are found on primary key
}
\description{
Counts any duplicate rows in a table when grouped by key(s)
}
