% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery.R
\name{bqAuth}
\alias{bqAuth}
\title{Wrapper for the set_service_token that
  uses BIGQUERY_ACCESS_TOKEN_PATH env var
  as default value for the secret token location.}
\usage{
bqAuth()
}
\description{
Required environment variables to use BigQuery helper functions:
  `BIGQUERY_PROJECT`` - name of the project in BigQuery.
  `BIGQUERY_DATASET`` - name of the default dataset in BigQuery.
  `BIGQUERY_ACCESS_TOKEN_PATH` - path to the json token file.
}
