% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery.R
\name{bqCopyTable}
\alias{bqCopyTable}
\title{Copies table in BigQuery}
\usage{
bqCopyTable(from, to)
}
\arguments{
\item{from}{name of the source table}

\item{to}{name of the desitnation table}
}
\value{
TRUE if the table has been succesfully copied
}
\description{
Copies table in BigQuery
}
