% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery.R
\name{bqCreatePartitionTable}
\alias{bqCreatePartitionTable}
\title{Creates partition table for a given sql}
\usage{
bqCreatePartitionTable(table, datasets, sql = NULL, file = NULL,
  existing.dates = NULL, missing.dates = NULL, priority = "INTERACTIVE")
}
\arguments{
\item{table}{name of the destination table}

\item{datasets}{list of Google Analytics properties to populate table for}

\item{sql}{sql to use a source of the data}

\item{file}{if sql is not provided it will be read from the file}

\item{existing.dates}{dates that should be skipped}

\item{missing.dates}{dates calculation for which will be enforced}

\item{priority}{priority of job execution. INTERACTIVE or BATCH.}
}
\description{
Parameters that will be passed to SQL

%1$s - is the name of the BigQuery dataset
%2$s - is the date (YYYYMMDD) of the partition
}
