% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery-partition.R
\name{bqCreatePartitionTable}
\alias{bqCreatePartitionTable}
\title{Creates partition table for a given sql}
\usage{
bqCreatePartitionTable(
  table,
  datasets,
  sql = NULL,
  file = NULL,
  existing.dates = NULL,
  missing.dates = NULL,
  priority = "INTERACTIVE",
  use.legacy.sql = bqUseLegacySql()
)
}
\arguments{
\item{table}{name of the destination table}

\item{datasets}{list of Google Analytics properties to populate table for}

\item{sql}{string with sql statement or query template}

\item{file}{if sql is not provided it will be read from the file}

\item{existing.dates}{dates that should be skipped}

\item{missing.dates}{dates calculation for which will be enforced}

\item{priority}{sets priority of job execution to INTERACTIVE or BATCH}

\item{use.legacy.sql}{switches SQL dialect.
Defaults to value set in `BIGQUERY_LEGACY_SQL` env.}
}
\description{
Parameters that will be passed to SQL for the placeholders:

%1$s - the name of the BigQuery dataset
%2$s - the date (YYYYMMDD) of the partition
}
