% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery-dataset.R
\name{bqDatasetExists}
\alias{bqDatasetExists}
\alias{bqCreateDataset}
\alias{bqDeleteDataset}
\alias{bqDatasetTables}
\title{Dataset manipulation functions}
\usage{
bqDatasetExists(dataset = bqDefaultDataset(),
  project = bqDefaultProject())

bqCreateDataset(dataset = bqDefaultDataset(),
  project = bqDefaultProject())

bqDeleteDataset(dataset = bqDefaultDataset(),
  project = bqDefaultProject(), delete.contents = FALSE)

bqDatasetTables(dataset = bqDefaultDataset(),
  project = bqDefaultProject())
}
\arguments{
\item{dataset}{name of the dataset}

\item{project}{name of the project}

\item{delete.contents}{removes all content from the dataset if is set to TRUE}
}
\value{
`bqDatasetExists()` - returns TRUE if dataset exists
}
\description{
Dataset manipulation functions

Gets list of tables for a given dataset
}
\details{
`bqDeleteDataset()` - You can protect dataset
  from programmatic deletion by adding `delete:never` label (key:value) to it.
}
