% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery.R
\name{bqDatasetTables}
\alias{bqDatasetTables}
\alias{bqDatasetExists}
\alias{bqCreateDataset}
\alias{bqDeleteDataset}
\title{Gets list of tables for a given dataset}
\usage{
bqDatasetTables(dataset = bqDefaultDataset(), project = bqDefaultProject())

bqDatasetExists(dataset = bqDefaultDataset(), project = bqDefaultProject())

bqCreateDataset(dataset = bqDefaultDataset(), project = bqDefaultProject())

bqDeleteDataset(dataset = bqDefaultDataset(), project = bqDefaultProject(),
  delete.contents = FALSE)
}
\arguments{
\item{dataset}{name of the dataset}

\item{project}{name of the project}

\item{delete.contents}{removes all content from the dataset if is set to TRUE}
}
\value{
`bqDatasetExists()` - returns TRUE if dataset exists
}
\description{
Gets list of tables for a given dataset

Dataset manipulation functions
}
\details{
`bqDeleteDataset()` - You can protect dataset from programmatic deletion by adding `delete:never` label (key:value) to it.
}
