% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery.R
\name{bqDeleteTable}
\alias{bqDeleteTable}
\title{Deletes table}
\usage{
bqDeleteTable(bq.table, bq.project = Sys.getenv("BIGQUERY_PROJECT"),
  bq.dataset = Sys.getenv("BIGQUERY_DATASET"))
}
\arguments{
\item{bq.table}{name of the table}

\item{bq.project}{name of the project}

\item{bq.dataset}{name of the dataset}
}
\value{
results of the execution from bigrquery::delete_table
}
\description{
Deletes table
}
