% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery.R
\name{bqDeleteTable}
\alias{bqDeleteTable}
\title{Deletes table}
\usage{
bqDeleteTable(table, dataset = Sys.getenv("BIGQUERY_DATASET"))
}
\arguments{
\item{table}{name of the table}

\item{dataset}{name of the dataset}
}
\value{
results of the execution from bigrquery::delete_table
}
\description{
Deletes table
}
