% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery.R
\name{bqExecuteSql}
\alias{bqExecuteSql}
\title{Gets data for a given SQL statement}
\usage{
bqExecuteSql(sql, ...)
}
\arguments{
\item{sql}{string with sql statment}

\item{...}{any parameters that will be used to fill in placeholders with sprintf}
}
\value{
results of execution as data.frame
}
\description{
Gets data for a given SQL statement
}
