% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery.R
\name{bqExtractTable}
\alias{bqExtractTable}
\title{Exports BigQuery table into Google Cloud Storage file}
\usage{
bqExtractTable(table, dataset = bqDefaultDataset(), compression = "GZIP")
}
\arguments{
\item{table}{name of the table to extract}

\item{dataset}{name of the dataset}

\item{compression}{The compression type to use for exported files. Possible
values include "GZIP", "DEFLATE", "SNAPPY", and "NONE". "DEFLATE" and
"SNAPPY" are only supported for Avro.}
}
\value{
object of `bq_job`
}
\description{
Exports BigQuery table into Google Cloud Storage file
}
\seealso{
?bigrquery::bq_table_save
}
