% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery.R
\name{bqInsertData}
\alias{bqInsertData}
\title{Inserts data into BigQuery table}
\usage{
bqInsertData(table, data, dataset = bqDefaultDataset(), append = TRUE,
  fields = NULL)
}
\arguments{
\item{table}{name of the target table}

\item{data}{data to be inserted}

\item{dataset}{name of the destination dataset}

\item{append}{specifies if data should be appended or truncated}

\item{fields}{list of fields with names and types (as bq_fields)}
}
\value{
results of execution
}
\description{
Inserts data into BigQuery table
}
