% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery.R
\name{bqInsertData}
\alias{bqInsertData}
\title{Inserts data into BigQuery table}
\usage{
bqInsertData(table, data, append = TRUE)
}
\arguments{
\item{table}{name of the target table}

\item{data}{data to be inserted}

\item{append}{specifies if data should be appended or truncated}
}
\value{
results of execution
}
\description{
Inserts data into BigQuery table
}
