% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery.R
\name{bqInsertLargeData}
\alias{bqInsertLargeData}
\title{Inserts LARGE data into BigQuery table}
\usage{
bqInsertLargeData(table, data, dataset = bqDefaultDataset(), chunks = 5,
  append = TRUE, schema.file = NULL)
}
\arguments{
\item{table}{name of the target table}

\item{data}{data to be inserted}

\item{dataset}{name of the destination dataset}

\item{chunks}{number of segments the data should be split into}

\item{append}{specifies if data should be appended or truncated}

\item{schema.file}{sets path to schema file for initialisation}
}
\value{
results of execution
}
\description{
Truncates the target table and appends data broken in chunks
}
