% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery.R
\name{bqQuery}
\alias{bqQuery}
\alias{bqGetData}
\alias{bqExecuteFile}
\alias{bqExecuteQuery}
\alias{bqExecuteSql}
\title{Functions that execute query against BigQuery database}
\usage{
bqGetData(sql = NULL, file = NULL, ...)

bqExecuteFile(file, ...)

bqExecuteQuery(query, ...)

bqExecuteSql(sql, ...)
}
\arguments{
\item{sql}{string with sql statment}

\item{file}{file with sql statment}

\item{...}{any parameters that will be used to fill in placeholders with sprintf}

\item{query}{template of the sql statement}
}
\value{
results of execution as data.table
}
\description{
Execute templated query given as string or file.
Placeholders in template are replaced with values provided in ellipsis parameter with sprintf.
}
