% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery.R
\name{bqGetData}
\alias{bqGetData}
\alias{bqExecuteFile}
\alias{bqExecuteQuery}
\alias{bqExecuteSql}
\title{Gets data for a given SQL statement or file that contains SQL}
\usage{
bqGetData(sql = NULL, file = NULL, ...)

bqExecuteFile(file, ...)

bqExecuteQuery(query, ...)

bqExecuteSql(sql, ...)
}
\arguments{
\item{sql}{sql statement}

\item{file}{file with sql statment}

\item{...}{any parameters that will be used to fill in placeholders with sprintf}

\item{query}{template of the sql statement}

\item{file}{file with sql statment}

\item{...}{any parameters that will be used to fill in placeholders with sprintf}

\item{sql}{string with sql statment}

\item{...}{any parameters that will be used to fill in placeholders with sprintf}
}
\value{
results of execution as data.frame

results of execution as data.frame

results of execution as data.frame
}
\description{
Gets data for a given SQL statement or file that contains SQL

Gets data for a given file that contains SQL statement

Gets data for a given SQL statement
}
