% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery-schema.R
\name{bqExtractSchema}
\alias{bqExtractSchema}
\alias{bqSaveSchema}
\title{Extract schema as JSON from data.table}
\usage{
bqExtractSchema(dt)

bqSaveSchema(dt, file)
}
\arguments{
\item{dt}{data.table, table name or bq_fields from which schema will be extracted into JSON}

\item{file}{path to the schema file}
}
\description{
Extract schema as JSON from data.table

Save file inferred from data.table to JSON file
}
