% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery-partition.R
\name{bqTransformPartition}
\alias{bqTransformPartition}
\title{Functions to transforms partitioned data form one table to another}
\usage{
bqTransformPartition(
  table,
  file,
  ...,
  missing.dates = NULL,
  priority = "INTERACTIVE",
  use.legacy.sql = bqUseLegacySql()
)
}
\arguments{
\item{table}{destination partition table where results of the query will be saved}

\item{file}{path to the sql file that will be used for the transformation}

\item{...}{parameters that will be passed via `sprintf` to build dynamic SQL.
partition date will be always passed first in format `yyyymmdd`
followed by arguments in `...`}

\item{missing.dates}{dates for which to run this function for}

\item{priority}{sets priority of job execution to INTERACTIVE or BATCH}

\item{use.legacy.sql}{Defaults to env variable if specified}
}
\description{
`bqTransformPartition` creates new partitions for the missing dates
}
