% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery.R
\name{bqUseLegacySql}
\alias{bqUseLegacySql}
\title{Gets the value from the corresponding environment variable as boolean
Determines which flavour of sql should be used by default.}
\usage{
bqUseLegacySql(x = NULL)
}
\arguments{
\item{x}{Sets `BIGQUERY_LEGACY_SQL` variable if set. Otherwise function returns value of the variable.}
}
\description{
Gets the value from the corresponding environment variable as boolean
Determines which flavour of sql should be used by default.
}
