% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{conformHeader}
\alias{conformHeader}
\title{Replace " ", "-" and "_" with a given separator in the header.}
\usage{
conformHeader(names, separator = ".")
}
\arguments{
\item{names}{column names to correct.}

\item{separator}{char to use between words in colnumn names.}
}
\description{
Replace " ", "-" and "_" with a given separator in the header.
}
