% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgresql.R
\name{dbExecuteQuery}
\alias{dbExecuteQuery}
\title{Executes SQL query agains the database}
\usage{
dbExecuteQuery(query)
}
\arguments{
\item{query}{string with sql query}
}
\description{
Executes SQL query agains the database
}
