% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgresql.R
\name{dbExecuteQueryFile}
\alias{dbExecuteQueryFile}
\title{Executes statement in the SQL file}
\usage{
dbExecuteQueryFile(file, ...)
}
\arguments{
\item{file}{path to the sql file}

\item{...}{any parameters that will be used
to fill in placeholders with sprintf}
}
\description{
Executes statement in the SQL file
}
