% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipes.R
\name{etl_read}
\alias{etl_read}
\title{Read (extract) part of ETL
reading tables}
\usage{
etl_read(from, ...)
}
\arguments{
\item{from}{DBI/R environment/file name}

\item{...}{args passed to read function}

\item{name}{character; table name}

\item{schema}{character;}

\item{asDT}{logical}

\item{lowercase}{logical;}
}
\description{
Read (extract) part of ETL
reading tables
}
