% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipes.R
\name{etl_write.odbc32}
\alias{etl_write.odbc32}
\title{Write tables via odbc32 connection}
\usage{
\method{etl_write}{odbc32}(to, x, name, schema = NULL,
  rownames = FALSE, safer = FALSE, ...)
}
\arguments{
\item{to}{}

\item{x}{data.table}

\item{name}{name of the target table}

\item{schema}{ignored}

\item{rownames}{logical;}

\item{safer}{logical;}

\item{...}{}
}
\description{
Write tables via odbc32 connection
}
