% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influx.R
\name{etlLogInflux}
\alias{etlLogInflux}
\title{Logs the result of a job in the influx metric table}
\usage{
etlLogInflux(con, db, job, val, metric, env)
}
\arguments{
\item{con}{a connection object}

\item{db}{the database to write}

\item{job}{the name of the job to log}

\item{val}{the variable to log}

\item{metric}{the metric to be written to}

\item{env}{the environment to write test/prod}
}
\description{
Logs the result of a job in the influx metric table
}
