% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery.R
\name{getExistingDates}
\alias{getExistingDates}
\title{Gets existing dates from wildcard tables in BigQuery}
\usage{
getExistingDates(dataset, table.prefix,
  project = Sys.getenv("BIGQUERY_PROJECT"))
}
\arguments{
\item{dataset}{name of a dataset}

\item{table.prefix}{name of the table before the wildcard}

\item{project}{name of the project}
}
\value{
string vector of dates
}
\description{
Gets existing dates from wildcard tables in BigQuery
}
