% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery.R
\name{getExistingDates}
\alias{getExistingDates}
\title{Gets existing dates from wildcard tables in BigQuery}
\usage{
getExistingDates(bq.dataset, table.prefix,
  bq.project = Sys.getenv("BIGQUERY_PROJECT"))
}
\arguments{
\item{bq.dataset}{name of a dataset}

\item{table.prefix}{name of the table before the wildcard}

\item{bq.project}{name of the project}
}
\value{
string vector of dates
}
\description{
Gets existing dates from wildcard tables in BigQuery
}
