% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigquery.R
\name{getLastID}
\alias{getLastID}
\title{Gets last id from a given field}
\usage{
getLastID(table, field, project = Sys.getenv("BIGQUERY_PROJECT"),
  dataset = Sys.getenv("BIGQUERY_DATASET"))
}
\arguments{
\item{table}{name of the table}

\item{field}{name of the field}

\item{project}{name of the project}

\item{dataset}{name of the dataset}
}
\value{
max value in a requested field
}
\description{
Gets last id from a given field
}
