% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlespreadsheets.R
\name{gsLoadAll}
\alias{gsLoadAll}
\title{Loads all sheets from a google spreadsheet into a list with the tab name as the list element name.}
\usage{
gsLoadAll(key, verbose = TRUE, lookup = TRUE, visibility = "private")
}
\arguments{
\item{key}{sheet-identifying information; a character vector of length one holding sheet title, key, browser URL or worksheets feed OR, in the case of gs_gs only, a googlesheet object}

\item{verbose}{OBSOLETE}

\item{lookup}{OBSOLETE}

\item{visibility}{OBSOLETE}
}
\description{
Loads all sheets from a google spreadsheet into a list with the tab name as the list element name.
}
