% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlespreadsheets.R
\name{gsLoadAll}
\alias{gsLoadAll}
\title{Loads all sheets from a google spreadsheet into a list with the tab name as the list element name.}
\usage{
gsLoadAll(key, token.file = "access_token.json")
}
\arguments{
\item{key}{Key of the google spreadsheet.}

\item{token.file}{Json access token file for Google API.}
}
\description{
Loads all sheets from a google spreadsheet into a list with the tab name as the list element name.
}
