% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlespreadsheets.R
\name{gsLoadAll}
\alias{gsLoadAll}
\title{Loads all sheets from a google spreadsheet into a list with the tab name as the list element name.}
\usage{
gsLoadAll(key)
}
\arguments{
\item{key}{Key of the google spreadsheet.}
}
\description{
Loads all sheets from a google spreadsheet into a list with the tab name as the list element name.
}
