% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlespreadsheets.R
\name{gsLoadSheet}
\alias{gsLoadSheet}
\title{Requires access_token.json file the working directory
Loads google spreadsheet via key}
\usage{
gsLoadSheet(key, tab, token.file = "access_token.json")
}
\arguments{
\item{key}{key of the google spreadsheet}

\item{token.file}{json access token file for Google API.}
}
\description{
Requires access_token.json file the working directory
Loads google spreadsheet via key
}
