% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlespreadsheets.R
\name{gsLoadSheet}
\alias{gsLoadSheet}
\title{Loads google spreadsheet via key}
\usage{
gsLoadSheet(key, tab)
}
\arguments{
\item{key}{key of the google spreadsheet}

\item{tab}{name of the tab from which data will be loaded}
}
\description{
Loads google spreadsheet via key
}
