% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tracking.R
\name{logModelPerformance}
\alias{logModelPerformance}
\title{Logs model performans data.frame to BigQuery table}
\usage{
logModelPerformance(data)
}
\arguments{
\item{data}{data with metrics of the model}
}
\description{
Logs model performans data.frame to BigQuery table
}
