% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{lowercase_names.data.table}
\alias{lowercase_names.data.table}
\alias{lowercase_names}
\title{Make all data.table names lowercase}
\usage{
\method{lowercase_names}{data.table}(DT, inplace = FALSE)

lowercase_names(x, ...)
}
\arguments{
\item{DT}{}
}
\description{
Make all data.table names lowercase
}
