% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-helpers.R
\name{readSql}
\alias{readSql}
\title{Creates SQL statement from the source file}
\usage{
readSql(file, ...)
}
\arguments{
\item{file}{file with sql statement template}

\item{...}{any parameters that will be used
to fill in placeholders in the template with sprintf}
}
\value{
SQL statement as a string
}
\description{
Creates SQL statement from the source file
}
