% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws-s3.R
\name{s3GetData}
\alias{s3GetData}
\title{Loads data from several files in S3 based on the path prefix}
\usage{
s3GetData(path, header = T, bucket = Sys.getenv("AWS_S3_BUCKET"),
  root = Sys.getenv("AWS_S3_ROOT"), s3Get.FUN = s3GetFile, fill = T)
}
\arguments{
\item{path}{is the path to the S3 object}

\item{header}{flag defines whether file has header}

\item{bucket}{name of the S3 bucket}

\item{root}{project root path that is appended before the path}

\item{s3Get.FUN}{Function that will be used to read data from the individual files.}

\item{fill}{see data.table::rbindList}
}
\value{
data.table with data combined from files.
}
\description{
Loads data from several files in S3 based on the path prefix
}
\note{
You need make sure that all targeted files have the same header signature:
  order of the fields should not matter.
}
