% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws-s3.R
\name{s3GetFile}
\alias{s3GetFile}
\title{Loads data from csv file in AWS S3}
\usage{
s3GetFile(path, header = T, bucket = Sys.getenv("AWS_S3_BUCKET"),
  root = Sys.getenv("AWS_S3_ROOT"))
}
\arguments{
\item{path}{is the path to the S3 object}

\item{header}{flag defines whether file has header}

\item{bucket}{name of the bucket}

\item{root}{project root path that is appended before the path}
}
\description{
Loads data from csv file in AWS S3
}

