% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws-s3.R
\name{s3PutFile}
\alias{s3PutFile}
\title{Saves a data.frame structure to AWS S3 as a csv file}
\usage{
s3PutFile(dt, path, bucket = Sys.getenv("AWS_S3_BUCKET"),
  root = Sys.getenv("AWS_S3_ROOT"))
}
\arguments{
\item{dt}{data.table to save}

\item{path}{S3 object path starting after root folder.}

\item{bucket}{name of the S3 bucket}

\item{root}{project root path that is appended before the path, e.g. "/prod/"}
}
\description{
Saves a data.frame structure to AWS S3 as a csv file
}
