% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws-s3.R
\name{s3PutFile}
\alias{s3PutFile}
\title{Saves the data frame to S3.}
\usage{
s3PutFile(df, file.name, path = "", bucket = Sys.getenv("AWS_S3_BUCKET"),
  root = Sys.getenv("AWS_S3_ROOT"))
}
\arguments{
\item{df}{data to save}

\item{file.name}{name of the file to be uploaded to S3.}

\item{path}{S3 path starting after root folder, excludes filename. example: "folder/"}

\item{bucket}{S3 bucket}

\item{root}{Root folder that will be appended before the path}
}
\description{
Saves the data frame to S3.
}
