% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws-s3.R
\name{s3PutFile}
\alias{s3PutFile}
\alias{s3PutFile.gz}
\alias{s3PutFile.rds}
\alias{s3PutFile.json.gz}
\title{Saves data to a file in AWS S3}
\usage{
s3PutFile(dt, path, bucket = Sys.getenv("AWS_S3_BUCKET"),
  root = Sys.getenv("AWS_S3_ROOT"), na.value = "")

s3PutFile.gz(dt, path, bucket = Sys.getenv("AWS_S3_BUCKET"),
  root = Sys.getenv("AWS_S3_ROOT"), na.value = "")

s3PutFile.rds(dt, path, bucket = Sys.getenv("AWS_S3_BUCKET"),
  root = Sys.getenv("AWS_S3_ROOT"))

s3PutFile.json.gz(dt, path, bucket = Sys.getenv("AWS_S3_BUCKET"),
  root = Sys.getenv("AWS_S3_ROOT"))
}
\arguments{
\item{dt}{data.table to save}

\item{path}{S3 object path starting after root folder.}

\item{bucket}{name of the S3 bucket}

\item{root}{project root path that is appended before the path, e.g. "/prod/"}

\item{na.value}{the string to use for missing values in the data}
}
\value{
TRUE if data was saved to S3.

`s3PutFile` saves data into csv file in S3

`s3PutFile.gz` saves data into `.gz` (csv archived) file in S3.

`s3PutFile.rds` saves data as `.rds` file in S3

`s3PutFile.json.gz` saves data as `.json.gz` (json archived) file in S3
}
\description{
Convention based wrapper functions that allow to save data in S3

Saves data.table as csv.gz to S3
}
