% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{safeLookup}
\alias{safeLookup}
\title{Function to safely append values to a data.table from another data.table via keys}
\usage{
safeLookup(data, lookup, by, select = setdiff(colnames(lookup), by))
}
\arguments{
\item{data}{data.table to which fields will be appended}

\item{lookup}{data.table from which fields will be appended}

\item{by}{fields that represent the keys}

\item{select}{fields that will be appended}
}
\value{
data set with new fields
}
\description{
Function to safely append values to a data.table from another data.table via keys
}
